moving the public facing download out of the lib package
