+++ library/src/main/java/com/novoda/downloadmanager/Download.java
@@ -1,4 +1,4 @@
-package com.novoda.downloadmanager.lib;
+package com.novoda.downloadmanager;
 
 public class Download {
